Refactor GETModerators
